working out the data loader and the data constants

+++ Question.java
@@ -3,9 +3,9 @@ import java.util.ArrayList;
 public class Question {
     private String question;
     private ArrayList<String> answers;
-    private Integer correctAnswer;
+    private Long correctAnswer;
 
-    public Question(String question, ArrayList<String> answers, Integer correctAnswers){
+    public Question(String question, ArrayList<String> answers, Long correctAnswers){
         this.question = question;
         this.answers = answers;
         this.correctAnswer = correctAnswers;
